requirements.txt
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
google-generativeai==0.3.2
python-dotenv==1.0.0
email-validator==2.1.0
colorama==0.4.6
beautifulsoup4==4.12.2
html2text==2020.1.16